% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{NOAA data frame}
}
\value{
seperate representation data
}
\description{
eq_clean_data
}
