% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data frame containing cleaned NOAA earthquake data}
}
\value{
HTML label string
}
\description{
eq_create_label
}
\examples{
\dontrun{eq_create_label("hello")}
}
