% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(v)
}
\arguments{
\item{v}{string to clean}
}
\value{
seperate representation of string
}
\description{
eq_location_clean
}
