% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{data}

\item{annot_col}{name of the HTML column}
}
\value{
A leaflet map with earthquakes and annotations.
}
\description{
eq_map
}
\examples{
\dontrun{eq_map(eq,"name")}
}
