% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{ggplot item}

\item{data}{ggplot item}

\item{stat}{ggplot item}

\item{position}{ggplot item}

\item{na.rm}{ggplot item}

\item{show.legend}{ggplot item}

\item{inherit.aes}{ggplot item}

\item{...}{ggplot item}
}
\value{
geom timeline label function
}
\description{
geom_timeline_label
}
\examples{
\dontrun{geom_timeline_label()}
}
